+++ jnn/camadas/Flatten.java
@@ -214,7 +214,7 @@ public class Flatten extends Camada implements Cloneable {
 	}
 
 	@Override
-	public double[] saidaParaArray() {
+	public float[] saidaParaArray() {
 		return saida().array();
 	}
